+++ consumer/consumer.go
@@ -13,8 +13,10 @@ import (
 	"github.com/nano-interactive/go-amqp/v2/serializer"
 )
 
-var _ io.Closer = Consumer[any]{}
-var _ io.Closer = (*Consumer[any])(nil)
+var (
+	_ io.Closer = Consumer[any]{}
+	_ io.Closer = (*Consumer[any])(nil)
+)
 
 type (
 	Message interface{}